requirements.txt
streamlit==1.39.0
pillow==10.4.0
numpy==1.26.4
pandas==2.2.3
matplotlib==3.9.2
scikit-learn==1.5.2
pytorch-grad-cam==1.5.4
jupyterlab==4.3.5
ipykernel==6.29.5
